working on #117
