+++ src/main/java/org/woehlke/twitterwall/frontend/controller/PagesController.java
@@ -1,29 +1,48 @@
-package org.woehlke.twitterwall.frontend.handler;
+package org.woehlke.twitterwall.frontend.controller;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
+import org.woehlke.twitterwall.frontend.controller.common.AbstractTwitterwallController;
+import org.woehlke.twitterwall.frontend.controller.common.Symbols;
+import org.woehlke.twitterwall.oodm.service.UserService;
 import org.woehlke.twitterwall.oodm.service.TaskService;
 import org.woehlke.twitterwall.scheduled.service.backend.TwitterApiService;
-import org.woehlke.twitterwall.frontend.common.AbstractTwitterwallController;
-import org.woehlke.twitterwall.frontend.common.Symbols;
-import org.woehlke.twitterwall.oodm.service.UserService;
 import org.woehlke.twitterwall.scheduled.service.persist.StoreOneTweet;
 import org.woehlke.twitterwall.scheduled.service.persist.StoreUserProfile;
 
-import javax.servlet.http.HttpServletRequest;
 
 /**
- * Created by tw on 17.06.17.
+ * Created by tw on 12.07.17.
  */
-@ControllerAdvice
-public class GlobalExceptionHandler extends AbstractTwitterwallController {
+@Controller
+public class PagesController extends AbstractTwitterwallController {
 
-    private static final Logger log = LoggerFactory.getLogger(GlobalExceptionHandler.class);
+    @RequestMapping("/")
+    public ModelAndView index(Model model) {
+        return new ModelAndView("redirect:/tweet/all");
+    }
+
+    @RequestMapping("/imprint")
+    public String imprint(Model model) {
+        log.info("-----------------------------------------");
+        logEnv();
+        String symbol = Symbols.IMPRINT.toString();
+        String title = "Impressum";
+        String subtitle = imprintSubtitle;
+        model = super.setupPage(model, title, subtitle, symbol);
+        String screenName = imprintScreenName;
+        super.addUserForScreenName(model,screenName);
+        log.info("-----------------------------------------");
+        return "imprint";
+    }
+
+    private static final Logger log = LoggerFactory.getLogger(PagesController.class);
 
     @Value("${twitterwall.frontend.menu.appname}")
     private String menuAppName;
@@ -43,6 +62,9 @@ public class GlobalExceptionHandler extends AbstractTwitterwallController {
     @Value("${twitterwall.frontend.imprint.screenName}")
     private String imprintScreenName;
 
+    @Value("${twitterwall.frontend.imprint.subtitle}")
+    private String imprintSubtitle;
+
     @Value("${twitterwall.frontend.idGoogleAnalytics}")
     private String idGoogleAnalytics;
 
@@ -52,36 +74,17 @@ public class GlobalExceptionHandler extends AbstractTwitterwallController {
 
     private final StoreUserProfile storeUserProfile;
 
-    private final UserService userService;
-
     private final TaskService taskService;
 
+    private final UserService userService;
+
     @Autowired
-    public GlobalExceptionHandler(TwitterApiService twitterApiService, StoreOneTweet storeOneTweet, StoreUserProfile storeUserProfile, UserService userService, TaskService taskService) {
+    public PagesController(TwitterApiService twitterApiService, StoreOneTweet storeOneTweet, StoreUserProfile storeUserProfile, TaskService taskService, UserService userService) {
         this.twitterApiService = twitterApiService;
         this.storeOneTweet = storeOneTweet;
         this.storeUserProfile = storeUserProfile;
-        this.userService = userService;
         this.taskService = taskService;
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ModelAndView handleIllegalArgumentException(HttpServletRequest request, Exception ex) {
-        log.warn("IllegalArgumentException occured :: URL=" + request.getRequestURL());
-        log.warn(ex.getMessage());
-        return getTemplate(request, ex);
-    }
-
-    private ModelAndView getTemplate(HttpServletRequest request, Exception ex) {
-        ModelAndView mav = new ModelAndView();
-        String symbol = Symbols.EXCEPTION.toString();
-        String title = "Exception";
-        String subtitle = ex.getMessage();
-        mav = super.setupPage(mav, title, subtitle, symbol);
-        mav.addObject("exception", ex);
-        mav.addObject("url", request.getRequestURL());
-        mav.setViewName("/exceptionhandler/persistentObjectNotFound");
-        return mav;
+        this.userService = userService;
     }
 
     @Override
