+++ src/main/java/org/woehlke/twitterwall/frontend/controller/common/UrlPaths.java
@@ -1,4 +1,4 @@
-package org.woehlke.twitterwall.frontend.common;
+package org.woehlke.twitterwall.frontend.controller.common;
 
 /**
  * Created by tw on 01.07.17.
