+++ src/main/java/org/woehlke/twitterwall/frontend/rest/MentionResource.java
@@ -1,4 +1,4 @@
-package org.woehlke.twitterwall.frontend.rest.entities;
+package org.woehlke.twitterwall.frontend.rest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -10,40 +10,40 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.woehlke.twitterwall.oodm.entities.Url;
-import org.woehlke.twitterwall.oodm.service.UrlService;
+import org.woehlke.twitterwall.oodm.entities.Mention;
+import org.woehlke.twitterwall.oodm.service.MentionService;
 
 
-import static org.woehlke.twitterwall.frontend.common.AbstractTwitterwallController.FIRST_PAGE_NUMBER;
+import static org.woehlke.twitterwall.frontend.controller.common.AbstractTwitterwallController.FIRST_PAGE_NUMBER;
 
 /**
  * Created by tw on 03.07.17.
  */
 @Controller
-@RequestMapping("/rest/url")
-public class UrlResource {
+@RequestMapping("/rest/mention")
+public class MentionResource {
 
     @RequestMapping(path="/count", method= RequestMethod.GET)
     public @ResponseBody
     long getCount() {
-    return this.urlService.count();
+    return this.mentionService.count();
     }
 
     @RequestMapping(path="/all", params = { "page" }, method= RequestMethod.GET)
     public @ResponseBody
-    Page<Url> getAll(@RequestParam(name= "page" ,defaultValue=""+FIRST_PAGE_NUMBER) int page) {
+    Page<Mention> getAll(@RequestParam(name= "page" ,defaultValue=""+FIRST_PAGE_NUMBER) int page) {
         Pageable pageRequest = new PageRequest(page, pageSize);
-        return this.urlService.getAll(pageRequest);
+        return this.mentionService.getAll(pageRequest);
     }
 
     @Value("${twitterwall.frontend.maxResults}")
     private int pageSize;
 
-    private final UrlService urlService;
+    private final MentionService mentionService;
 
     @Autowired
-    public UrlResource(UrlService urlService) {
-        this.urlService = urlService;
+    public MentionResource(MentionService mentionService) {
+        this.mentionService = mentionService;
     }
 
 }
