+++ src/main/java/org/woehlke/twitterwall/frontend/rest/TickerSymbolResource.java
@@ -1,4 +1,4 @@
-package org.woehlke.twitterwall.frontend.rest.entities;
+package org.woehlke.twitterwall.frontend.rest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -10,40 +10,39 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.woehlke.twitterwall.oodm.entities.Mention;
-import org.woehlke.twitterwall.oodm.service.MentionService;
+import org.woehlke.twitterwall.oodm.entities.TickerSymbol;
+import org.woehlke.twitterwall.oodm.service.TickerSymbolService;
 
 
-import static org.woehlke.twitterwall.frontend.common.AbstractTwitterwallController.FIRST_PAGE_NUMBER;
+import static org.woehlke.twitterwall.frontend.controller.common.AbstractTwitterwallController.FIRST_PAGE_NUMBER;
 
 /**
  * Created by tw on 03.07.17.
  */
 @Controller
-@RequestMapping("/rest/mention")
-public class MentionResource {
+@RequestMapping("/rest/tickersymbol")
+public class TickerSymbolResource {
 
     @RequestMapping(path="/count", method= RequestMethod.GET)
     public @ResponseBody
     long getCount() {
-    return this.mentionService.count();
+    return this.tickerSymbolService.count();
     }
 
     @RequestMapping(path="/all", params = { "page" }, method= RequestMethod.GET)
     public @ResponseBody
-    Page<Mention> getAll(@RequestParam(name= "page" ,defaultValue=""+FIRST_PAGE_NUMBER) int page) {
+    Page<TickerSymbol> getAll(@RequestParam(name= "page" ,defaultValue=""+FIRST_PAGE_NUMBER) int page) {
         Pageable pageRequest = new PageRequest(page, pageSize);
-        return this.mentionService.getAll(pageRequest);
+        return this.tickerSymbolService.getAll(pageRequest);
     }
 
     @Value("${twitterwall.frontend.maxResults}")
     private int pageSize;
 
-    private final MentionService mentionService;
+    private final TickerSymbolService tickerSymbolService;
 
     @Autowired
-    public MentionResource(MentionService mentionService) {
-        this.mentionService = mentionService;
+    public TickerSymbolResource(TickerSymbolService tickerSymbolService) {
+        this.tickerSymbolService = tickerSymbolService;
     }
-
 }
