+++ src/main/java/org/woehlke/twitterwall/frontend/rest/TweetResource.java
@@ -13,9 +13,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.woehlke.twitterwall.oodm.entities.Tweet;
 import org.woehlke.twitterwall.oodm.service.TweetService;
 
-import java.util.List;
-
-import static org.woehlke.twitterwall.frontend.common.AbstractTwitterwallController.FIRST_PAGE_NUMBER;
+import static org.woehlke.twitterwall.frontend.controller.common.AbstractTwitterwallController.FIRST_PAGE_NUMBER;
 
 /**
  * Created by tw on 03.07.17.
