+++ src/main/java/org/woehlke/twitterwall/frontend/rest/UrlCacheResource.java
@@ -1,4 +1,4 @@
-package org.woehlke.twitterwall.frontend.rest.entities;
+package org.woehlke.twitterwall.frontend.rest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -10,39 +10,40 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.woehlke.twitterwall.oodm.entities.Media;
-import org.woehlke.twitterwall.oodm.service.MediaService;
+import org.woehlke.twitterwall.oodm.entities.UrlCache;
+import org.woehlke.twitterwall.oodm.service.UrlCacheService;
 
-import static org.woehlke.twitterwall.frontend.common.AbstractTwitterwallController.FIRST_PAGE_NUMBER;
+import static org.woehlke.twitterwall.frontend.controller.common.AbstractTwitterwallController.FIRST_PAGE_NUMBER;
 
 /**
  * Created by tw on 03.07.17.
  */
 @Controller
-@RequestMapping("/rest/media")
-public class MediaResource {
+@RequestMapping("/rest/urlcache")
+public class UrlCacheResource {
 
     @RequestMapping(path="/count", method= RequestMethod.GET)
     public @ResponseBody
     long getCount() {
-    return this.mediaService.count();
+    return this.urlCacheService.count();
     }
 
     @RequestMapping(path="/all", params = { "page" }, method= RequestMethod.GET)
     public @ResponseBody
-    Page<Media> getAll(@RequestParam(name= "page" ,defaultValue=""+FIRST_PAGE_NUMBER) int page) {
+    Page<UrlCache> getAll(@RequestParam(name= "page" ,defaultValue=""+FIRST_PAGE_NUMBER) int page) {
         Pageable pageRequest = new PageRequest(page, pageSize);
-        return this.mediaService.getAll(pageRequest);
+        return this.urlCacheService.getAll(pageRequest);
     }
 
     @Value("${twitterwall.frontend.maxResults}")
     private int pageSize;
 
-    private final MediaService mediaService;
+    private final UrlCacheService urlCacheService;;
 
     @Autowired
-    public MediaResource(MediaService mediaService) {
-        this.mediaService = mediaService;
+    public UrlCacheResource(UrlCacheService urlCacheService) {
+        this.urlCacheService = urlCacheService;
     }
+
 
 }
