+++ src/main/java/org/woehlke/twitterwall/frontend/rest/UrlResource.java
@@ -1,4 +1,4 @@
-package org.woehlke.twitterwall.frontend.rest.entities;
+package org.woehlke.twitterwall.frontend.rest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -10,39 +10,40 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.woehlke.twitterwall.oodm.entities.TickerSymbol;
-import org.woehlke.twitterwall.oodm.service.TickerSymbolService;
+import org.woehlke.twitterwall.oodm.entities.Url;
+import org.woehlke.twitterwall.oodm.service.UrlService;
 
 
-import static org.woehlke.twitterwall.frontend.common.AbstractTwitterwallController.FIRST_PAGE_NUMBER;
+import static org.woehlke.twitterwall.frontend.controller.common.AbstractTwitterwallController.FIRST_PAGE_NUMBER;
 
 /**
  * Created by tw on 03.07.17.
  */
 @Controller
-@RequestMapping("/rest/tickersymbol")
-public class TickerSymbolResource {
+@RequestMapping("/rest/url")
+public class UrlResource {
 
     @RequestMapping(path="/count", method= RequestMethod.GET)
     public @ResponseBody
     long getCount() {
-    return this.tickerSymbolService.count();
+    return this.urlService.count();
     }
 
     @RequestMapping(path="/all", params = { "page" }, method= RequestMethod.GET)
     public @ResponseBody
-    Page<TickerSymbol> getAll(@RequestParam(name= "page" ,defaultValue=""+FIRST_PAGE_NUMBER) int page) {
+    Page<Url> getAll(@RequestParam(name= "page" ,defaultValue=""+FIRST_PAGE_NUMBER) int page) {
         Pageable pageRequest = new PageRequest(page, pageSize);
-        return this.tickerSymbolService.getAll(pageRequest);
+        return this.urlService.getAll(pageRequest);
     }
 
     @Value("${twitterwall.frontend.maxResults}")
     private int pageSize;
 
-    private final TickerSymbolService tickerSymbolService;
+    private final UrlService urlService;
 
     @Autowired
-    public TickerSymbolResource(TickerSymbolService tickerSymbolService) {
-        this.tickerSymbolService = tickerSymbolService;
+    public UrlResource(UrlService urlService) {
+        this.urlService = urlService;
     }
+
 }
