+++ src/main/java/org/woehlke/twitterwall/oodm/dao/impl/TaskDaoImpl.java
@@ -6,7 +6,6 @@ import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Repository;
 import org.woehlke.twitterwall.oodm.entities.Task;
 import org.woehlke.twitterwall.oodm.dao.TaskDao;
-import org.woehlke.twitterwall.oodm.dao.parts.impl.DomainDaoImpl;
 
 import javax.persistence.TypedQuery;
