+++ src/main/java/org/woehlke/twitterwall/oodm/dao/impl/TickerSymbolDaoImpl.java
@@ -5,7 +5,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Repository;
 import org.woehlke.twitterwall.oodm.entities.TickerSymbol;
-import org.woehlke.twitterwall.oodm.dao.parts.impl.DomainDaoImpl;
 import org.woehlke.twitterwall.oodm.dao.TickerSymbolDao;
 
 import javax.persistence.EntityManager;
